spelling corrections, cleared test space, seperated pr
